split plugin into separate 'modules'
