+++ src/declarations.d.ts
@@ -1,0 +1,1 @@
+declare module 'pagerank.js'
